% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-names.R
\name{excel_names}
\alias{excel_names}
\title{List defined named names in a workbook}
\usage{
excel_names(path)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}
}
\description{
Lists all defined names in a workbook, with any associated local range references.
The list of defined names is especially useful when you want to avoid hard-coding
the range of cells to be read from a workbook. Instead, if a range of cells to be
read is named in the workbook, this function allows you to obtain the actual cell
range, which may then be used in a subsequent call to \link{read_excel} to read the
range.
}
\details{
The function returns a \code{tibble} with three columns:
\itemize{
\item name: the name defined in the workbook
\item context: for names defined on a worksheet, the worksheet name; NA for workbook names
\item ref: for names defined as a range in the same workbook, the associated range
}
}
\examples{
names <- excel_names(readxl_example("names.xlsx"))
ref <- names[names$name == "data"]$ref
read_excel(readxl_example("names.xlsx"), range = ref)
}
