% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/named-ranges.R
\name{named_ranges}
\alias{named_ranges}
\title{List all the named ranges in an excel spreadsheet.}
\usage{
named_ranges(path)
}
\arguments{
\item{path}{Path to the xls/xlsx file}
}
\description{
List all the named ranges in an excel spreadsheet.
}

