% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{readxl_example}
\alias{readxl_example}
\title{Get path to readxl example}
\usage{
readxl_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
readxl comes bundled with some example files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\examples{
readxl_example()
readxl_example("datasets.xlsx")
}
