% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{readxl_progress}
\alias{readxl_progress}
\title{Determine whether to show progress spinner}
\usage{
readxl_progress()
}
\description{
By default, readxl displays a progress spinner \strong{unless}
one of the following is \code{TRUE}:
\itemize{
\item The spinner is explicitly disabled by setting
\code{options(readxl.show_progress = FALSE)}.
\item The code is run in a non-interactive session (\code{interactive()} is \code{FALSE}).
\item The code is run by knitr / rmarkdown.
\item The code is run in an RStudio notebook chunk.
readxl uses the \href{https://cran.r-project.org/package=progress}{progress package} under-the-hood and
therefore is also sensitive to any options that it consults.
}
}
