% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zip_buffer.R
\name{zip_buffer}
\alias{zip_buffer}
\title{Load a file inside a zip into a raw vector}
\usage{
zip_buffer(zip_path, file_path)
}
\description{
Load a file inside a zip into a raw vector
}
\keyword{internal}

