% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-format.R
\name{excel_format}
\alias{excel_format}
\title{Determine file format}
\usage{
excel_format(path, guess = TRUE)
}
\arguments{
\item{path}{Path to the xls/xlsx file}

\item{guess}{Logical. Whether to guess format based on the file itself, if
the extension is neither \code{"xlsx"} nor \code{"xls"}.}
}
\value{
Character vector with values \code{"xlsx"}, \code{"xls"}, or \code{NA}.
}
\description{
Determine if files are xlsx or xls. First the file extension is consulted. If
that is unsuccessful and \code{guess = TRUE} and the file exists, the format is
guessed from the \href{https://en.wikipedia.org/wiki/List_of_file_signatures}{filesignature} or "magic
number".
}
\examples{
files <- c(
  "a.xlsx",
  "b.xls",
  "c.png",
  file.path(R.home("doc"), "html", "logo.jpg"),
  readxl_example("clippy.xlsx"),
  readxl_example("deaths.xls")
)
excel_format(files)
}
