% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readxl-package.R
\docType{package}
\name{readxl-package}
\alias{readxl}
\alias{readxl-package}
\title{readxl: Read Excel Files}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Import excel files into R. Supports '.xls' via the embedded 'libxls' C library \url{https://github.com/libxls/libxls} and '.xlsx' via the embedded 'RapidXML' C++ library \url{http://rapidxml.sourceforge.net}. Works on Windows, Mac and Linux without external dependencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://readxl.tidyverse.org}
  \item \url{https://github.com/tidyverse/readxl}
  \item Report bugs at \url{https://github.com/tidyverse/readxl/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@posit.co} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@posit.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})
}

Other contributors:
\itemize{
  \item Posit, PBC (Copyright holder of all R code and all C/C++ code without explicit copyright attribution) [copyright holder, funder]
  \item Marcin Kalicinski (Author of included RapidXML code) [contributor, copyright holder]
  \item Komarov Valery (Author of included libxls code) [contributor, copyright holder]
  \item Christophe Leitienne (Author of included libxls code) [contributor, copyright holder]
  \item Bob Colbert (Author of included libxls code) [contributor, copyright holder]
  \item David Hoerl (Author of included libxls code) [contributor, copyright holder]
  \item Evan Miller (Author of included libxls code) [contributor, copyright holder]
}

}
\keyword{internal}
