% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readxl.R
\docType{package}
\name{readxl-package}
\alias{readxl}
\alias{readxl-package}
\title{readxl: Read Excel Files}
\description{
Import excel files into R. Supports '.xls' via the embedded
'libxls' C library <https://sourceforge.net/projects/libxls/> and '.xlsx' via
the embedded 'RapidXML' C++ library <https://rapidxml.sourceforge.net>. Works on
Windows, Mac and Linux without external dependencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidyverse/readxl}
  \item Report bugs at \url{https://github.com/tidyverse/readxl/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@rstudio.com}

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio (Copyright holder of
      all R code and all C/C++ code without explicit copyright attribution) [copyright holder, funder]
  \item Marcin Kalicinski (Author of included RapidXML code) [contributor, copyright holder]
  \item Komarov Valery (Author of included libxls code) [contributor, copyright holder]
  \item Christophe Leitienne (Author of included libxls code) [contributor, copyright holder]
  \item Bob Colbert (Author of included libxls code) [contributor, copyright holder]
  \item David Hoerl (Author of included libxls code) [contributor, copyright holder]
}

}
\keyword{internal}
